Fix Linting Issues
